+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 - PIM-2874: Bad title on failed submit
 - PIM-3836: Fix translations of a custom job label instance
+- PIM-3909: Keep channel filter between product datagrid and edit form
 
 # 1.3.4 (2015-03-11)